Accommodate Kinetis oddball alignment requiremets
TX buffers submitted to the Kinetis ENET controller need to be 8B
aligned.  This is a longer alignment than would otherwise normally be
necessary on this 32-bit system.

lwip has existing facilities to align the payload in a frame, but it
didn't have facilities to ensure alignment of the entire frame.
Introduce such alignment mechanisms.

This behavior is enabled by specifying MEM_FRAME_ALIGNMENT in
lwipopts.h.  It can be used at the same time as MEM_ALIGNMENT so long
as the two can be simultaneously satisifed.

Note that it's possible to specify frame and payload alignment
requirements that are incompatible.  This will result in a failed
assertion at runtime.