Hack to 8B align DHCP packet buffers
Adding a UDP header to an 8B aligned pbuf allocated with
PBUF_LAYER_TRANSPORT results in a 4B but not 8B aligned header which
the Kinetis can't handle.  This breaks DHCP which builds its pbuf in
place including adding headers.

Just have the DHCP consume 4 bytes of header space up front so that the
result is properly 8B aligned in the end.  I mean, it works, right?